shfmt and shellcheck

+++ tools/cp_sample_worlds.bash
@@ -2,52 +2,50 @@
 
 set -eu
 
-USAGE="$(basename $0) <World Dir>"
+USAGE="$(basename "$0") <World Dir>"
 
-project_root=$(cd "$(dirname $0)/.."; pwd)
+project_root=$(
+  cd "$(dirname "$0")/.."
+  pwd
+)
 dest="${project_root}/docs/sample_world"
 
 copy_files=(
-    biomes.png
-    main.ttw
-    map_info.xml
-    prefabs.xml
-    radiation.png
-    spawnpoints.xml
-    splat3_processed.png
-    splat4_processed.png
+  biomes.png
+  main.ttw
+  map_info.xml
+  prefabs.xml
+  radiation.png
+  spawnpoints.xml
+  splat3_processed.png
+  splat4_processed.png
 )
 
-main() {
-    if [[ $# -ne 1 ]]
-    then
-        echo $USAGE
-        exit 1
-    fi
+if [[ $# -ne 1 ]]; then
+  echo "$USAGE"
+  exit 1
+fi
 
-    cd "${project_root}"
+cd "${project_root}"
 
-    src="$1"
+src="$1"
 
-    echo "Src:  $src"
-    echo "Dest: $dest"
+echo "Src:  $src"
+echo "Dest: $dest"
 
-    set -x
+set -x
 
-    rm -frv "$dest/"*
+rm -frv "${dest:?}/"*
 
-    (
-        cd "$src"
-        cp -v "${copy_files[@]}" "$dest"
-    )
+(
+  cd "$src"
+  cp -v "${copy_files[@]}" "$dest"
+)
 
-    map_width="$(identify -format '%[width]' "$src/splat3_processed.png")"
-    if [[ -z "${map_width}" ]]; then
-        echo "Empty map width"
-        exit 1
-    fi
-    echo "MapWidth: ${map_width}"
-    npx tsx ./tools/generate-dtm-png.ts "$src/dtm.raw" "${map_width}" "$dest/dtm.png"
-}
-
-main "$@"
+map_width="$(identify -format '%[width]' "$src/splat3_processed.png")"
+if [[ -z "${map_width}" ]]; then
+  echo "Empty map width"
+  exit 1
+fi
+echo "MapWidth: ${map_width}"
+npx tsx ./tools/generate-dtm-png.ts "$src/dtm.raw" "${map_width}" "$dest/dtm.png"
